+++ app/src/main/java/org/simple/clinic/storage/migrations/RoomMigrationsModule.kt
@@ -108,7 +108,8 @@ class RoomMigrationsModule {
       migration100: Migration_100,
       migration101: Migration_101,
       migration102: Migration_102,
-      migration103: Migration_103
+      migration103: Migration_103,
+      migration104: Migration_104
   ): List<Migration> {
     return listOf(
         migration_3_4,
@@ -210,7 +211,8 @@ class RoomMigrationsModule {
         migration100,
         migration101,
         migration102,
-        migration103
+        migration103,
+        migration104
     )
   }
 }