+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Handle user interactions for overdue search results
 - Bump Sentry to v6.1.2
 - Consume `OverdueSearchViewEffect` in `OverdueSearchEffectHandler`
+- Implement query to search overdue patients
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300